update audio playing capabilitys to signals current state (based on 80deb301e54c766acd3bfc7edbec400b03a44c4d). Fixes #1088

+++ src/org/thoughtcrime/securesms/ConversationItem.java
@@ -355,7 +355,7 @@ public class ConversationItem extends LinearLayout
       public void onStop() {}
 
       @Override
-      public void onProgress(double progress, int millis) {}
+      public void onProgress(double progress, long millis) {}
 
       @Override
       public void onReceivedDuration(int millis) {
